CatchUpResponse now includes the currentTerm of the sender.
This should now mean that masters own a term across all slots; previously it
would have been possible for the master to reconfigure to a disjoint set of
voting nodes, which then catch-up, but do not sync up on terms. They could then
elect their own master in an earlier term (but later slot) than before. This
isn't a safety concern but it simplifies debugging if masters own terms regardless
of the current slot.